chore(mpt): remove old version
