+++ crates/executor/client/src/io.rs
@@ -2,7 +2,7 @@ use std::iter::once;
 
 use eyre::Result;
 use itertools::Itertools;
-use openvm_mpt::EthereumState2;
+use openvm_mpt::EthereumState;
 use openvm_witness_db::WitnessDb;
 use reth_primitives::{Block, Header, TransactionSigned};
 use reth_trie::TrieAccount;
@@ -29,7 +29,7 @@ pub struct ClientExecutorInput {
     #[serde_as(as = "Vec<alloy_consensus::serde_bincode_compat::Header>")]
     pub ancestor_headers: Vec<Header>,
     /// Network state as of the parent block.
-    pub parent_state: EthereumState2,
+    pub parent_state: EthereumState,
     /// Requests to account state and storage slots.
     pub state_requests: HashMap<Address, Vec<U256>>,
     /// Account bytecodes.
@@ -51,7 +51,7 @@ impl ClientExecutorInput {
 
 impl WitnessInput for ClientExecutorInput {
     #[inline(always)]
-    fn state(&self) -> &EthereumState2 {
+    fn state(&self) -> &EthereumState {
         &self.parent_state
     }
 
@@ -79,7 +79,7 @@ impl WitnessInput for ClientExecutorInput {
 /// A trait for constructing [`WitnessDb`].
 pub trait WitnessInput {
     /// Gets a reference to the state from which account info and storage slots are loaded.
-    fn state(&self) -> &EthereumState2;
+    fn state(&self) -> &EthereumState;
 
     /// Gets the state trie root hash that the state referenced by
     /// [state()](trait.WitnessInput#tymethod.state) must conform to.
